inventory_management/requirements.txt
asgiref==3.9.1
dj-database-url==3.0.1
Django==5.2.5
django-filter==25.1
djangorestframework==3.16.1
python-decouple==3.8
sqlparse==0.5.3
tzdata==2025.2
